# Conflicts:
#	stroom-dashboard/stroom-dashboard-client/src/main/java/stroom/dashboard/client/table/DownloadViewImpl.java
#	stroom-dashboard/stroom-dashboard-client/src/main/java/stroom/dashboard/client/table/TablePresenter.java
#	stroom-dashboard/stroom-dashboard-client/src/main/resources/stroom/dashboard/client/table/DownloadViewImpl.ui.xml
#	stroom-dashboard/stroom-dashboard-impl/src/main/java/stroom/dashboard/impl/DashboardServiceImpl.java

+++ stroom-core-client-widget/src/main/java/edu/ycp/cs/dh/acegwt/client/ace/AceEditor.java
@@ -872,7 +872,7 @@ public class AceEditor extends Composite implements RequiresResize, HasText, Tak
 		editor.getSession().setValue(output);
 	}-*/;
 
-    public native void setScrollMargin(int top, int bottom, int left, int right) /*-{
+    public native void setScrollMargin(double top, double bottom, double left, double right) /*-{
 		var editor = this.@edu.ycp.cs.dh.acegwt.client.ace.AceEditor::editor;
 		editor.renderer.setScrollMargin(top, bottom, left, right);
 	}-*/;
